projet-scrapping-docker/requirements_vector.txt
langchain==0.0.267
pinecone-client==2.2.2
openai==0.28.1
tqdm==4.66.1
